requirements.txt
torch>=2.0.0
numpy>=1.21.0
scipy>=1.7.0
wandb>=0.15.0
scikit-learn>=1.0.0
PyYAML>=6.0
pickle5>=0.0.11
jaxtyping>=0.2.0
